Move filterOptions into tenant store and fix store provider refs
