+++ src/RootStore/TenantStore/filterOptions.js
@@ -25,8 +25,8 @@ import {
   VIOLATION_TYPE,
   LEVEL_1_SUPERVISION_LOCATION,
   LEVEL_2_SUPERVISION_LOCATION,
-} from "../../../constants/filterTypes";
-import { US_MO, US_PA } from "../../../RootStore/TenantStore/lanternTenants";
+} from "../../constants/filterTypes";
+import { US_MO, US_PA } from "./lanternTenants";
 
 const METRIC_PERIODS = {
   options: [
