+++ src/RootStore/__tests__/FiltersStore.test.js
@@ -22,7 +22,7 @@ import FiltersStore from "../FiltersStore";
 import getTenantMappings, { tenantMappings } from "../TenantStore/tenants";
 import { LANTERN_TENANTS } from "../TenantStore/lanternTenants";
 
-jest.mock("../../StoreProvider");
+jest.mock("../../components/StoreProvider");
 jest.mock("../TenantStore/tenants");
 
 let rootStore;
