+++ src/RootStore/__tests__/TenantStore.test.js
@@ -25,7 +25,7 @@ import { METADATA_NAMESPACE } from "../../constants";
 
 jest.mock("@auth0/auth0-spa-js");
 jest.mock("../utils/user");
-jest.mock("../../StoreProvider");
+jest.mock("../../components/StoreProvider");
 jest.mock("../../api/metrics/metricsClient");
 jest.mock("../DataStore/DataStore");
