+++ src/utils/downloads/getFilters.js
@@ -2,7 +2,7 @@ import pipe from "lodash/fp/pipe";
 import find from "lodash/fp/find";
 import get from "lodash/fp/get";
 import { humanReadableTitleCase } from "../transforms/labels";
-import { SUPERVISION_LEVELS } from "../../core/tenants/constants/filterOptions";
+import { SUPERVISION_LEVELS } from "../../RootStore/TenantStore/filterOptions";
 
 function formatMetricPeriodMonthsFilter(metricPeriodMonths) {
   switch (metricPeriodMonths) {
